Forgot to make new branch, committing changes to make new branch

+++ scripts/data_extraction_cleanup.py
@@ -18,8 +18,7 @@ import pandas as pd
 from sklearn.model_selection import train_test_split
 import helper_functions as hf
 
-# Confirm we are in the correct directory, otherwise break script 
-# and prompt user to move to correct directory
+# Confirm the correct directory; break script and prompt user to move to correct directory otherwise
 filepath = os.getcwd()
 if not filepath.endswith('myersBriggsNLPAnalysis'):
 	print('\nYou do not appear to be in the correct directory,\
@@ -33,8 +32,6 @@ wide_data = 'data/mbti_wide.csv'
 token_data = 'data/mbti_tokenized.csv'
 clean_data = 'data/mbti_cleaned.csv'
 columns = np.array(['type', 'posts'])
-raw_df = pd.read_csv(raw_data, header = 0)
-wide_df = pd.read_csv(wide_data, header = 0)
 
 ################
 # Tokenize data 
@@ -49,22 +46,34 @@ This is data in 'long' format; my justification is here:
 https://sejdemyr.github.io/r-tutorials/basics/wide-and-long/
 
 And tokenize the line as well (keeping hashtags, urls, etc.)
-I used the regex package, but the nltk.tokenize package has many different tokenizers available
+# I used the regex package, but the nltk.tokenize package has many different tokenizers
 https://www.nltk.org/_modules/nltk/tokenize/regexp.html
 Check helper_function for more detail on tokenization method used
 '''
 
-start = time.time()
-token_df = hf.parallelize(func = hf.tokenize_data, df = raw_df)
-end = time.time()
-elapsed = end - start
-print('Time elapsed: %.2f' % elapsed)
+# Check if tokenized file exists; if not, create it
+# If it does, ask user if they want to make it again
+file_exists = os.path.isfile(token_data)
+if file_exists:
+    prompt_user = input('\nThe tokenized data appears to already exist.\n' +
+        'Would you like to generate it again? (Y/n) ')
+if prompt_user == 'n':
+    print('\nContinuing to clean data without stopwords.')
+elif (prompt_user == 'Y') or (not file_exists):
+    start = time.time()
+    token_df = hf.parallelize(func = hf.tokenize_data, df = raw_df)
+    end = time.time()
+    elapsed = end - start
+    print('Time elapsed: %.2f' % elapsed)
+    print('Tokenized Data Shape:', token_df.shape)
+    print('Head of Tokenized Data:', token_df.head(10))
 
-print('Tokenized Data Shape:', token_df.shape)
-print('Head of Tokenized Data:', token_df.head(10))
+    # Write to csv
+    token_df.to_csv(token_data, columns = list(token_df.columns.values), index = False)
 
-# Write to csv
-token_df.to_csv(token_data, columns = list(token_df.columns.values), index = False)
+else:
+    print('\nInvalid input, please run script again with valid input to create tokenized data.')
+    sys.exit(1)
 
 ##################
 # Remove stopwords
@@ -78,18 +87,28 @@ Articles, conjections and some adverbs are all candidates for this (i.e. 'to', '
 I will be analyzing the data with stopwords included, but would also like a copy of data without stopwords
 For analysis as well as use for the machine learning models I will be implementing
 '''
-start = time.time()
-#clean_df = hf.parallelize(func = hf.tokenize_data(remove_stopwords = True), df = raw_df)
-clean_df = hf.tokenize_data(raw_df, filter_stopwords = True)
-end = time.time()
-elapsed = end - start
-print('Time elapsed: %.2f' % elapsed)
+file_exists = os.path.isfile(clean_data)
+if file_exists:
+    prompt_user = input('\nThe cleaned data appears to already exist.\n' +
+        'Would you like to generate it again? (Y/n) ')
+if (prompt_user == 'Y') or (not file_exists):
+    start = time.time()
+    #clean_df = hf.parallelize(func = hf.tokenize_data(remove_stopwords = True), df = raw_df)
+    clean_df = hf.tokenize_data(raw_df, filter_stopwords = True)
+    end = time.time()
+    elapsed = end - start
+    print('Time elapsed: %.2f' % elapsed)
+    print('Cleaned Data Shape:', clean_df.shape)
+    print('Head of Cleaned Data:', clean_df.head(10))
 
-print(clean_df.shape)
-print(clean_df.head(10))
-
-# Write to csv
-clean_df.to_csv(clean_data, columns = list(clean_df.columns.values), index = False)
+    # Write to csv
+    clean_df.to_csv(clean_data, columns = list(clean_df.columns.values), index = False)
+elif prompt_user == 'n':
+    print('\nContinuing through rest of script.')
+else:
+    print('\nInvalid input, please run script again with a valid input\
+    to create the cleaned data.')
+    sys.exit(1)
 
 ##################################################
 # Make different versions of our data for analysis
@@ -107,20 +126,21 @@ We will now subset the data into various parts to be used in the other scripts
 '''
 
 # Declare different processed and unprocessed objects for further analysis
+raw_df = pd.read_csv(raw_data, header = 0)
 raw_type = raw_df['type']
 raw_posts = raw_df['posts']
 
-# Split wide data (every user has a line, can have multiple tweets)
+wide_df = pd.read_csv(wide_data, header = 0)
 wide_type = wide_df['type']
 wide_posts = wide_df['posts']
 
-# Split token data (each tweet has an individual line)
+token_df = pd.read_csv(token_data, header = 0)
 token_type = token_df['type']
 token_posts = token_df['posts']
 
-# Split cleaned (no stopwords) data
-cleaned_type = cleaned_df['type']
-cleaned_posts = cleaned_df['posts']
+clean_df = pd.read_csv(clean_data, header = 0)
+clean_type = clean_df['type']
+clean_posts = clean_df['posts']
 
 # Split up data into training and testing datasets
 # To evaluate effectiveness of model training
